move: 이미지 관련 아이콘을 images/ 로 이동 및 관련 코드 수정
