+++ src/pages/mypage/index.tsx
@@ -4,8 +4,8 @@ import HeaderWithLogo from '@components/HeaderWithLogo';
 import UserIcon from '@assets/icons/user.svg';
 import GoIcon from '@assets/icons/go.svg';
 import ExitIcon from '@assets/icons/exit.svg';
-import MyFundingIcon from '@assets/icons/my-funding.svg';
-import ParticipatedFundingIcon from '@assets/icons/participated_funding.svg';
+import MyFundingIcon from '@assets/images/my-funding.svg';
+import ParticipatedFundingIcon from '@assets/images/participated_funding.svg';
 import React, { useEffect } from 'react';
 import Image from 'next/image';
 import { useRouter } from 'next/router';
